+++ packages/firebase_auth/test/firebase_auth_test.dart
@@ -44,6 +44,9 @@ void main() {
           case "updateProfile":
             return null;
             break;
+          case "updateEmail":
+            return null;
+            break;
           case "fetchProvidersForEmail":
             return new List<String>(0);
             break;
@@ -320,6 +323,25 @@ void main() {
       );
     });
 
+    test('delete', () async {
+      final FirebaseUser user = await auth.currentUser();
+      await user.delete();
+
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall(
+            'currentUser',
+            arguments: null,
+          ),
+          isMethodCall(
+            'delete',
+            arguments: null,
+          ),
+        ],
+      );
+    });
+
     test('sendPasswordResetEmail', () async {
       await auth.sendPasswordResetEmail(
         email: kMockEmail,
@@ -352,6 +374,20 @@ void main() {
           },
         ),
       ]);
+    });
+
+    test('updateEmail', () async {
+      final String updatedEmail = 'atestemail@gmail.com';
+      auth.updateEmail(email: updatedEmail);
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall(
+            'updateEmail',
+            arguments: <String, String>{'email': updatedEmail},
+          ),
+        ],
+      );
     });
 
     test('signInWithCustomToken', () async {
@@ -437,11 +473,11 @@ void main() {
 }
 
 Map<String, dynamic> mockFirebaseUser(
-        {String providerId: kMockProviderId,
-        String uid: kMockUid,
-        String displayName: kMockDisplayName,
-        String photoUrl: kMockPhotoUrl,
-        String email: kMockEmail}) =>
+        {String providerId = kMockProviderId,
+        String uid = kMockUid,
+        String displayName = kMockDisplayName,
+        String photoUrl = kMockPhotoUrl,
+        String email = kMockEmail}) =>
     <String, dynamic>{
       'isAnonymous': true,
       'isEmailVerified': false,